Split the sign-up destination predicate into its own util
One util per file, so each filename describes what it exports.
