+++ packages/twenty-server/src/engine/core-modules/auth/utils/has-provisioned-sign-up-destination.util.ts
@@ -1,0 +1,13 @@
+import {
+  isWorkspaceProvisioned,
+  type WorkspaceActivationStatus,
+} from 'twenty-shared/workspace';
+
+export const hasProvisionedSignUpDestination = (
+  availableWorkspacesForSignUp: {
+    workspace: { activationStatus: WorkspaceActivationStatus };
+  }[],
+): boolean =>
+  availableWorkspacesForSignUp.some(({ workspace }) =>
+    isWorkspaceProvisioned(workspace),
+  );
